+++ Packages/com.unity.inputsystem/CHANGELOG.md
@@ -60,6 +60,11 @@ This release includes a number of Quality-of-Life improvements for a range of co
 
 ### Changed
 
+- The logic for resetting devices on focus loss has changed somewhat:
+  * When focus is lost, all devices are forcibly reset to their default state. As before, a `RequestResetCommand` for each device is also sent to the backend but regardless of whether the device responds or not, the input state for the device will be overwritten to default.
+  * __Noisy controls are exempted from resets__. The assumption here is that noisy controls most often represent sensor readings of some kind (e.g. tracking data) and snapping the values back to their default will usually
+  * If `Application.runInBackground` is `true`, all devices that return `true` from `InputDevice.canRunInBackground` are exempted from resets entirely. This, for example, allows XR devices to continue running regardless of focus change.
+  * This fixes problems such as keyboard keys getting stuck when alt-tabbing between applications (case 1206199).
 - `InputControlExtensions.GetStatePtrFromStateEvent` no longer throws `InvalidOperationException` when the state format for the event does not match that of the device. It simply returns `null` instead (same as when control is found in the event's state).
 
 #### Actions
@@ -86,6 +91,8 @@ This release includes a number of Quality-of-Life improvements for a range of co
   * This usually manifested itself as large accumulated mouse deltas leading to such effects as the camera immediately jerking around on game start.
 - Removing a device no longer has the potential of corrupting state change monitors (and thus actions getting triggered) from other devices.
   * This bug led to input being missed on a device once another device had been removed.
+- Noise masks for controls were initialized incorrectly, leading to incorrect input noise detection.
+- `TrackedDevice` layout is no longer incorrectly registered as `Tracked Device`.
 
 ## [1.0.0-preview.3] - 2019-11-14